# perf: add critical database indices for eval_results table