## Summary

This PR adds critical missing database indices that dramatically improve eval page loading performance.

## Problem

The  table (260k+ rows) was missing a composite index on , which is the primary access pattern for eval pages. This caused full table scans on every page load.

## Solution

Added the following indices:
-  - Composite index on (eval_id, test_idx) 
-  - For filtering by pass/fail
-  - For filtering by error type
-  - For pagination with timestamps

## Performance Results

**Before:**
- Basic eval query: ~319ms
- Filtered queries: 500-1000ms

**After:**
- Basic eval query: ~0.79ms (99.8% improvement)
- Filtered queries: ~0.48ms
- Search queries: ~0.53ms

## Testing

1. Benchmarked with production-size database (3.7GB, 260k eval results)
2. Verified query execution plans show index usage
3. Tested all query patterns (basic, filtered, search, pagination)

## Migration

To apply:
```bash
npm run migration:indices up
```

To rollback:
```bash
npm run migration:indices down
```

## Next Steps

- Fix SQL injection vulnerability in queryTestIndices
- Implement caching for eval summaries (still takes ~600ms)